refactor(bindings): address iOS write-watchdog review — shared clock, local reset, scope docs
Follow-up cleanups from PR review of the iOS relay write-stall watchdog:

- Extract a shared MonotonicClock helper (Foundation-only, like
  SocketGenerationTracker) and route both InternetManager.monotonicNowMs
  and OfflineProtocolModule's background-duration gate through it. Removes
  the hand-copied mach_continuous_time timebase+math that lived in two
  files and could silently drift, breaking the shared-clock assumption the
  foreground gate and the watchdog both rely on.

- Reset the write-stall watchdog in connect() so "a fresh generation
  starts with an empty FIFO" is a local guarantee, independent of every
  teardown path having reset() first. stalledAgeMs reads the head
  regardless of generation, so a stray abandoned entry surviving into a new
  generation could false-stall the healthy socket; resetting on connect (on
  messageQueue, before the post-auth poll arms any write) closes that window.

- Document the watchdog's coverage scope: only poll-path data/control writes
  are watched (the ones that pin inFlightControlPrimaries). An idle
  foreground zombie with an empty outbox self-heals on the next send or via
  the foreground forceReconnect belt — spelled out so the scope isn't
  over-read as full socket-liveness detection.

Register MonotonicClock.swift in Package.swift and MeshSdk.podspec.
swift test: 105 passing.
